Integrate the state machine methods into the state machine definition.
This makes the code easier to read.
